+++ src/main/java/hansung/mannayo/mannayoserverapplication/Model/Entity/Report.java
@@ -17,7 +17,7 @@ import java.time.LocalTime;
 public class Report {
 
     @Id @GeneratedValue
-    private Integer idReport;
+    private Long idReport;
 
     @NotNull
     @Enumerated(EnumType.STRING)
