+++ src/main/java/hansung/mannayo/mannayoserverapplication/Repository/JjimRepository.java
@@ -3,5 +3,5 @@ package hansung.mannayo.mannayoserverapplication.Repository;
 import hansung.mannayo.mannayoserverapplication.Model.Entity.Jjim;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JjimRepository extends JpaRepository<Jjim, Integer> {
+public interface JjimRepository extends JpaRepository<Jjim, Long> {
 }
